+++ src/helpers/functions/Game.js
@@ -3,3 +3,5 @@ const { gameStatuses, gameDefaultOptions } = require("../constants/Game");
 exports.getGameStatuses = () => JSON.parse(JSON.stringify(gameStatuses));
 
 exports.getGameDefaultOptions = () => JSON.parse(JSON.stringify(gameDefaultOptions));
+
+exports.getGameCardById = (game, cardId) => game.cards.find(({ _id }) => _id.toString() === cardId.toString());